pyproject.toml
[project]
name = "x-remove"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.2",
    "lxml>=5.3.0",
    "fake-useragent>=1.5.1",
    "x-remove",
    "uvicorn[standard]>=0.30.6",
    "quart>=0.19.6",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/x_remove"]

[tool.uv.sources]
x-remove = { workspace = true }
